[version 0.2.4] Added unit test coverage
I'm incrementing version to close out the unit test issue.

https://github.com/kuhl-haus/kuhl-haus-metrics/issues/3

The tests aren't perfect but good enough for now.  I don't like pushing a bunch of dev commits to the mainline branch but all of the CI workflows trigger on mainline so I created another issue to add a workflow for CI on the development branch.

https://github.com/kuhl-haus/kuhl-haus-metrics/issues/5